Refactor launcher
- first working version
- RCB now uses new launcher classes
